refactor: clean up settings dialog tabs
- rename Members tab to Users (Members.vue -> MembersSettings.vue) and
  replace EmojiSettings with CustomEmojiSettings
- drop the redundant per-tab SettingsPanel; the shell already wraps each
  tab in its own SettingsPanel/TabsContent
- fix settings overlay cold-load via loadRouteLocation + shallowRef route
- fix typescript errors across the tabs: Select v-model bindings, useCall
  param types, useDoc name getters, nullable profile doc
